pyproject.toml
[tool.poetry]
name = "qenerate"
version = "0.6.3"
description = "Code Generator for GraphQL Query and Fragment Data Classes"
authors = ["Red Hat - Service Delivery - AppSRE <sd-app-sre@redhat.com>"]
maintainers = ["Red Hat - Service Delivery - AppSRE <sd-app-sre@redhat.com>"]
license = "Apache-2.0"
homepage = "https://github.com/app-sre/qenerate"
repository = "https://github.com/app-sre/qenerate"
documentation = "https://github.com/app-sre/qenerate"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.22"
graphql-core = "^3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
requests-mock = "^1.9"
types-requests = "^2.27"
pyfakefs = "^4.5"
mypy = "^1.8.0"
ruff = "^0.1.11"

[tool.poetry.scripts]
qenerate = 'qenerate.cli:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = 'py39'
required-version = "0.1.11"

src = ["qenerate"]
extend-exclude = ["demo", ".local"]
fix = true

[tool.ruff.lint]
preview = true

# defaults are ["E4", "E7", "E9", "F"]
extend-select = [
    # flake8 default rules
    "E1", # preview rule
    "E2", # preview rule
    "W",
    # isort
    "I",
    # pylint
    "PL",
]
ignore = [
    "PLR0913", # Too many arguments
    "PLR0917", # Too many positional arguments
    "PLR6301", # Method `...` could be a function, class method, or static method
]
[tool.ruff.format]
preview = true

[tool.ruff.lint.isort]
known-first-party = ["qenerate"]
